pyproject.toml
[project]
name = "litestar-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litestar[standard]>=2.13.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.10.3",
    "tortoise-orm>=0.22.1",
]
